enhance:系统功能和UI优化

+++ backend/src/main/java/com/review/agent/repository/DefaultLlmProviderRepository.java
@@ -3,5 +3,5 @@ package com.review.agent.repository;
 import com.review.agent.entity.pojo.DefaultLlmProvider;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DefaultLlmProviderRepository extends JpaRepository<DefaultLlmProvider, Long> {
+public interface DefaultLlmProviderRepository extends JpaRepository<DefaultLlmProvider, Integer> {
 }
